+++ packages/asc-ui/src/components/deprecated/IconButton/IconButton.tsx
@@ -2,14 +2,18 @@ import React from 'react'
 import IconButtonStyle, {
   Props as IconButtonStyleProps,
 } from './IconButtonStyle'
-import { Icon } from '../..'
+import Icon from '../../Icon'
 
 type Props = {
   onClick?: React.MouseEventHandler<HTMLButtonElement>
-} & IconButtonStyleProps
+} & IconButtonStyleProps &
+  React.ButtonHTMLAttributes<HTMLButtonElement>
 
+/**
+ * @deprecated Use the Button component in combination with the Icon component
+ */
 const IconButton: React.FC<Props> = ({ children, onClick, ...otherProps }) => (
-  <IconButtonStyle onClick={onClick} {...otherProps} square>
+  <IconButtonStyle onClick={onClick} {...otherProps} square variant="blank">
     <Icon>{children}</Icon>
   </IconButtonStyle>
 )