+++ CHANGELOG.md
@@ -13,11 +13,19 @@ Prefix the change with one of these keywords:
 - *Security*: in case of vulnerabilities.
 
 ## [Unreleased]
+- Added: react-docgen-typescript-loader to storybook config for showing props with description
+- Added: font-display: swap; in font-face definitions
+- Added: seperate tsconfig for building the lib: now excludes .test and .stories files
+- Deprecated: Made IconButton, ButtonBar and ShareBar deprecated. Also refactored this in other 
+  components than use IconButton
+- Added: stories for buttons
+- Changed: Refactored the Button component to use the prop "variants" instead of "color". 
+- Removed: ShareBar, named export however still exists
 
 ## [0.9.2] - 2019-07-24
 ### ASC-UI
 - Added: Blog Components
-- Change: *BREAKING* Heading: `as` prop is now `$as`
+- Changed: *BREAKING* Heading: `as` prop is now `$as`
 
 ## [0.9.0] - 2019-07-12
 ### ASC-UI